refactor: move processing event class to separate file
