+++ Sources/RudderStackAnalytics/EventQueue/EventQueue.swift
@@ -70,27 +70,3 @@ extension EventQueue {
         self.eventWriter?.stop()
     }
 }
-
-// MARK: - ProcessingEvent
-/**
- * ProcessingEvent represents an event in the processing pipeline.
- * It can be either a regular message event or a flush signal.
- */
-class ProcessingEvent {
-    var type: ProcessingEventType
-    var event: Event?
-    
-    init(type: ProcessingEventType, event: Event? = nil) {
-        self.type = type
-        self.event = event
-    }
-}
-
-// MARK: - ProcessingEventType
-/**
- * Enum representing the different types of processing events.
- */
-enum ProcessingEventType {
-    case message
-    case flush
-}
